+++ src/components/stats/UserBrowsers.tsx
@@ -1,7 +1,7 @@
+import { useUserBrowsers } from '@/hooks/useUserBrowsers';
 import { BarChartTemplate } from '../charts/BarChartTemplate';
 import { Card } from '../common/Card';
 import { Title } from '../common/Title';
-import { useUserBrowsers } from '../hook/useUserBrowsers';
 
 export const UserBrowsers = () => {
   const { data } = useUserBrowsers();
