+++ src/components/stats/UserCountries.tsx
@@ -1,7 +1,7 @@
+import { useUserCountries } from '@/hooks/useUserCountries';
 import { PieChartTemplate } from '../charts/PieChartTemplate';
 import { Card } from '../common/Card';
 import { Title } from '../common/Title';
-import { useUserCountries } from '../hook/useUserCountries';
 
 export const UserCountries = () => {
   const { data } = useUserCountries();
