+++ src/hooks/useUserCountries.ts
@@ -1,13 +1,13 @@
 import { fetchData } from '@/utils/api';
 import { useSuspenseQuery } from '@tanstack/react-query';
-import { PieChartDataType } from '../charts/PieChartTemplate';
+import { PieChartDataType } from '../components/charts/PieChartTemplate';
 
 interface UserCountriesDataType {
   country: string;
   count: number;
 }
 
-export const useUserCounties = () => {
+export const useUserCountries = () => {
   return useSuspenseQuery<PieChartDataType[]>({
     queryKey: ['userCountries'],
     queryFn: async () => {
